Refactor Photos platform plugin structure
