+++ mobile/apps/photos/plugins/ente_photos_platform/ios/Flutter/PhotosPlatformPlugin.swift
@@ -1,0 +1,20 @@
+@preconcurrency import Flutter
+import Foundation
+
+@MainActor
+public final class PhotosPlatformPlugin: NSObject, @preconcurrency FlutterPlugin {
+    private let deviceHealthAdapter: DeviceHealthChannelAdapter
+
+    private init(registrar: FlutterPluginRegistrar) {
+        deviceHealthAdapter = DeviceHealthChannelAdapter(registrar: registrar)
+        super.init()
+    }
+
+    public static func register(with registrar: FlutterPluginRegistrar) {
+        registrar.publish(PhotosPlatformPlugin(registrar: registrar))
+    }
+
+    public func detachFromEngine(for registrar: FlutterPluginRegistrar) {
+        deviceHealthAdapter.detach()
+    }
+}
