+++ mobile/apps/photos/plugins/ente_photos_platform/ios/Sources/PhotosPlatformCore/DeviceHealthModels.swift
