+++ mobile/apps/photos/plugins/ente_photos_platform/ios/ente_photos_platform.podspec
@@ -6,7 +6,7 @@ Pod::Spec.new do |s|
   s.license          = { :type => 'AGPL-3.0-only' }
   s.author           = { 'Ente' => 'code@ente.com' }
   s.source           = { :path => '.' }
-  s.source_files     = 'Classes/**/*'
+  s.source_files     = '{Flutter,Sources}/**/*'
   s.dependency 'Flutter'
   s.platform         = :ios, '15.1'
   s.swift_version    = '5.0'
